@@ -1,6 +1,6 @@
 {
-    "name": "rwoverdijk/assetmanager",
-    "description": "An assetmanager module for Laminas.",
+    "name": "dcs/assetmanager",
+    "description": "An asset manager module for Laminas.",
     "license": "BSD-2-Clause",
     "type": "library",
     "keywords": [
@@ -21,24 +21,24 @@
             "homepage": "http://ocramius.github.com/"
         }
     ],
+    "replace": {
+        "rwoverdijk/assetmanager": "self.version"
+    },
     "require": {
-        "php": "^7.3 || ^8.0",
-        "laminas/laminas-modulemanager": "^2.7.2",
+        "php": "^8.1",
         "laminas/laminas-stdlib": "^3.2.1",
-        "laminas/laminas-servicemanager": "^2.7.6 || ^3.1.1",
-        "laminas/laminas-loader": "^2.5.1",
-        "laminas/laminas-eventmanager": "^2.6.3 || ^3.0.1",
-        "laminas/laminas-mvc": "^2.7.9 || ^3.0.2",
-        "assetic/framework": "^2.0 || ^3.0",
+        "laminas/laminas-servicemanager": "^3.1.1",
+        "laminas/laminas-eventmanager": "^3.0.1",
+        "assetic/framework": "^3.0",
         "laminas/laminas-dependency-plugin": "^2.0"
     },
     "require-dev": {
         "laminas/laminas-http": "^2.5.4",
-        "laminas/laminas-view": "^2.8.1",
-        "laminas/laminas-cache": "^2.7.1",
-        "laminas/laminas-console": "^2.6",
+        "laminas/laminas-cache": "^3.0",
+        "laminas/laminas-cache-storage-adapter-blackhole": "^2.0",
+        "laminas/laminas-cache-storage-adapter-memory": "^2.0",
         "squizlabs/php_codesniffer": "^3.5.6",
-        "phpunit/phpunit": "^8.5.8 || ^9.1"
+        "phpunit/phpunit": "^9.1"
     },
     "suggest": {
         "laminas/laminas-mvc-console": "laminas-mvc-console provides the ability to expose laminas-mvc as a console application"
