@@ -4,35 +4,18 @@ namespace AssetManager\Service;
 
 use AssetManager\Resolver\CollectionResolver;
 use Interop\Container\ContainerInterface;
-use Laminas\ServiceManager\FactoryInterface;
-use Laminas\ServiceManager\ServiceLocatorInterface;
 
-class CollectionResolverServiceFactory implements FactoryInterface
+class CollectionResolverServiceFactory implements \Laminas\ServiceManager\Factory\FactoryInterface
 {
-    /**
-     * @inheritDoc
-     */
-    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
+    public function __invoke(ContainerInterface $container, $requestedName, array $options = null): CollectionResolver
     {
         $config      = $container->get('config');
-        $collections = array();
+        $collections = [];
 
         if (isset($config['asset_manager']['resolver_configs']['collections'])) {
             $collections = $config['asset_manager']['resolver_configs']['collections'];
         }
 
-        $collectionResolver = new CollectionResolver($collections);
-
-        return $collectionResolver;
-    }
-
-    /**
-     * {@inheritDoc}
-     *
-     * @return CollectionResolver
-     */
-    public function createService(ServiceLocatorInterface $serviceLocator)
-    {
-        return $this($serviceLocator, CollectionResolver::class);
+        return new CollectionResolver($collections);
     }
 }
