@@ -4,10 +4,7 @@ namespace AssetManager\Service;
 
 use AssetManager\Resolver\MapResolver;
 use Interop\Container\ContainerInterface;
-use Interop\Container\Exception\ContainerException;
-use Laminas\ServiceManager\Exception\ServiceNotCreatedException;
-use Laminas\ServiceManager\Exception\ServiceNotFoundException;
-use Laminas\ServiceManager\FactoryInterface;
+use Laminas\ServiceManager\Factory\FactoryInterface;
 use Laminas\ServiceManager\ServiceLocatorInterface;
 
 class MapResolverServiceFactory implements FactoryInterface
@@ -15,26 +12,22 @@ class MapResolverServiceFactory implements FactoryInterface
     /**
      * @inheritDoc
      */
-    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
+    public function __invoke(ContainerInterface $container, $requestedName, array $options = null): MapResolver
     {
         $config = $container->get('config');
-        $map    = array();
+        $map    = [];
 
         if (isset($config['asset_manager']['resolver_configs']['map'])) {
             $map = $config['asset_manager']['resolver_configs']['map'];
         }
 
-        $patchStackResolver = new MapResolver($map);
-
-        return $patchStackResolver;
+        return new MapResolver($map);
     }
 
     /**
      * {@inheritDoc}
-     *
-     * @return MapResolver
      */
-    public function createService(ServiceLocatorInterface $serviceLocator)
+    public function createService(ServiceLocatorInterface $serviceLocator): MapResolver
     {
         return $this($serviceLocator, MapResolver::class);
     }
