@@ -2,14 +2,16 @@
 
 namespace AssetManager\Resolver;
 
+use Assetic\Contracts\Asset\AssetInterface;
+
 interface ResolverInterface
 {
     /**
      * Resolve an Asset
      *
-     * @param   string  $path   The path to resolve.
+     * @param string $path The path to resolve.
      *
-     * @return  \Assetic\Contracts\Asset\AssetInterface|null Asset instance when found, null when not.
+     * @return  AssetInterface|null Asset instance when found, null when not.
      */
     public function resolve($path);
 }
