@@ -5,10 +5,7 @@ namespace AssetManager\Service;
 use AssetManager\Resolver\AliasPathStackResolver;
 use AssetManager\Resolver\PathStackResolver;
 use Interop\Container\ContainerInterface;
-use Interop\Container\Exception\ContainerException;
-use Laminas\ServiceManager\Exception\ServiceNotCreatedException;
-use Laminas\ServiceManager\Exception\ServiceNotFoundException;
-use Laminas\ServiceManager\FactoryInterface;
+use Laminas\ServiceManager\Factory\FactoryInterface;
 use Laminas\ServiceManager\ServiceLocatorInterface;
 
 class AliasPathStackResolverServiceFactory implements FactoryInterface
