@@ -4,32 +4,22 @@ namespace AssetManager\Service;
 
 use AssetManager\Resolver\PrioritizedPathsResolver;
 use Interop\Container\ContainerInterface;
-use Laminas\ServiceManager\FactoryInterface;
+use Laminas\ServiceManager\Factory\FactoryInterface;
 use Laminas\ServiceManager\ServiceLocatorInterface;
 
 class PrioritizedPathsResolverServiceFactory implements FactoryInterface
 {
-    /**
-     * @inheritDoc
-     */
-    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
+    public function __invoke(ContainerInterface $container, $requestedName, array $options = null): PrioritizedPathsResolver
     {
         $config                   = $container->get('config');
         $prioritizedPathsResolver = new PrioritizedPathsResolver();
-        $paths                    = isset($config['asset_manager']['resolver_configs']['prioritized_paths'])
-            ? $config['asset_manager']['resolver_configs']['prioritized_paths']
-            : array();
+        $paths                    = $config['asset_manager']['resolver_configs']['prioritized_paths'] ?? [];
         $prioritizedPathsResolver->addPaths($paths);
 
         return $prioritizedPathsResolver;
     }
 
-    /**
-     * {@inheritDoc}
-     *
-     * @return PrioritizedPathsResolver
-     */
-    public function createService(ServiceLocatorInterface $serviceLocator)
+    public function createService(ServiceLocatorInterface $serviceLocator): PrioritizedPathsResolver
     {
         return $this($serviceLocator, PrioritizedPathsResolver::class);
     }
